Cleanup and lottie upgrade
